Address warnings for integer expressions of different signedness
Address warnings for comparisons of integer expressions of different
signedness, for example:

../firehose.c:384:31: warning: comparison of integer expressions of
   different signedness: ‘int’ and ‘long unsigned int’ [-Wsign-compare]
  384 |                 for (i = 0; i < ARRAY_SIZE(sector_sizes); i++) {

In all places, where signed value is casted to unsigned (size_t for
instance), there is always explicitly handling of possible negative
value beforehand

+++ vip.c
@@ -395,7 +395,7 @@ int vip_transfer_init(struct qdl_device *qdl, const char *vip_table_path)
 
 out_cleanup:
 	close(qdl->vip_data.signed_table_fd);
-	for (int i = 0; i < qdl->vip_data.chained_num - 1; ++i)
+	for (size_t i = 0; i < qdl->vip_data.chained_num - 1; ++i)
 		close(qdl->vip_data.chained_fds[i]);
 	return -1;
 }
@@ -403,7 +403,7 @@ out_cleanup:
 void vip_transfer_deinit(struct qdl_device *qdl)
 {
 	close(qdl->vip_data.signed_table_fd);
-	for (int i = 0; i < qdl->vip_data.chained_num - 1; ++i)
+	for (size_t i = 0; i < qdl->vip_data.chained_num - 1; ++i)
 		close(qdl->vip_data.chained_fds[i]);
 }
